News website
This news blog project is a CMS based project which has two phases first is a frontend phase and the second is the backend in frontend we will have an actual design of the web page like how will header look like where will the menu how many post pages will be there where will be the footer this all thing comes in frontend page and coming to the backend where the backend is password protected where the owner will access backend pages by adding its username and password he can access all the backend pages it may also call as control panel so a question comes what owner can do? an owner can add a new page without knowing any type of programming language we will make this system where the owner can create a new page the owner can change the content of the webpage can change a logo of the web page can change the theme of webpages can create a user in the webpage this is also called dynamic website
In this new blog website, we will see a different category-wise author-wise news and he can search particular news on how to build a login and logout page how to used session we will see how to upload images on server and database and pagination.